requirements.txt
Flask==2.3.2
tensorflow==2.19.0
pillow==9.5.0
python-multipart==0.0.5
numpy==1.26.4
flask-cors==3.0.10
openai
python-dotenv
Flask-Session==0.5.0